fix(app): extract debug civilian spawn to separate library
CI flutter analyze failed with bool-not-a-type cascades from the large
app_event_handler_scope unit; moving applyDebugCivilianSpawnAtCapital and
helpers to app_event_handler_debug_spawn.dart restores clean parsing.
